oracle/requirements.txt
bitcoin==1.1.42
certifi==2017.11.5
chardet==3.0.4
click==6.7
cytoolz==0.9.0
eth-abi==0.5.0
eth-keys==0.1.0b4
eth-tester==0.1.0b11
eth-utils==0.7.4
ethereum==1.0.8
ethjsonrpc==0.3.0
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lru-dict==1.1.6
MarkupSafe==1.0
pbkdf2==1.3
py-solc==1.4.0
pycryptodome==3.4.7
pyethash==0.1.27
pysha3==1.0.2
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
rlp==0.6.0
scrypt==0.8.0
semantic-version==2.6.0
toolz==0.9.0
typing==3.6.4
urllib3==1.22
web3==2.4.0
Werkzeug==0.14.1
